# services/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_workived_services_pkg_apperr.AppError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
...
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.LoginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/internal_auth.User'
    type: object
  internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.RefreshResponse:
    properties:
      access_token:
        type: string
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  internal_auth.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_login_at:
        type: string
    type: object
  internal_auth.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  internal_department.CreateDepartmentRequest:
    properties:
      name:
        maxLength: 150
        minLength: 1
        type: string
      parent_id:
        type: string
    required:
    - name
    type: object
  internal_department.Department:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      organisation_id:
        type: string
      parent_id:
        type: string
      updated_at:
        type: string
    type: object
  internal_department.UpdateDepartmentRequest:
    properties:
      name:
        maxLength: 150
        minLength: 1
        type: string
      parent_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HR and operations superapp for SMB founders in Indonesia and UAE.
  title: Workived API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_auth.LoginResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      summary: Log in and obtain an access token
      tags:
      - Auth
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Reads the refresh token from the httpOnly cookie. Falls back to
        request body.
      parameters:
      - description: Refresh token (only if not using cookie)
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_auth.RefreshResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      summary: Refresh the access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_auth.User'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      summary: Register a new user account
      tags:
      - Auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      summary: Verify a user's email address
      tags:
      - Auth
  /departments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_department.Department'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: List departments
      tags:
      - Departments
    post:
      consumes:
      - application/json
      parameters:
      - description: Department details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_department.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_department.Department'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Create a new department
      tags:
      - Departments
  /departments/{id}:
    delete:
      parameters:
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate (soft-delete) a department
      tags:
      - Departments
    put:
      consumes:
      - application/json
      parameters:
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_department.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_department.Department'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                $ref: '#/definitions/github_com_workived_services_pkg_apperr.AppError'
            type: object
      security:
      - BearerAuth: []
      summary: Update a department
      tags:
      - Departments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"